controllers message, space, booking - modifed + db small fixes, exception handling

+++ Backend/src/main/java/com/example/backend/service/impl/SpaceServiceImpl.java
@@ -1,14 +1,12 @@
 package com.example.backend.service.impl;
 
-import com.example.backend.dtos.Space.AddSpaceRequest;
-import com.example.backend.dtos.Space.EditSpaceRequest;
-import com.example.backend.dtos.Space.SpaceFilter;
-import com.example.backend.dtos.Space.SpaceResponse;
+import com.example.backend.dtos.Space.*;
 import com.example.backend.entity.Booking;
 import com.example.backend.entity.Space;
 import com.example.backend.entity.User;
 import com.example.backend.enums.Availibility;
 import com.example.backend.exception.ResourceNotFoundException;
+import com.example.backend.exception.UnauthorizedException;
 import com.example.backend.mapper.ObjectMapper;
 import com.example.backend.repository.BookingRepository;
 import com.example.backend.repository.SpaceRepository;
@@ -42,8 +40,15 @@ public class SpaceServiceImpl implements SpaceService {
         if(space == null){
             throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Space not correctly converted");
         }
-        spaceRepository.save(space);
-        return ObjectMapper.mapSpaceToSpaceResponse(space);
+        Space result = spaceRepository.save(space);
+        if(result == null){
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Space not saved");
+        }
+        SpaceResponse spaceResponse = ObjectMapper.mapSpaceToSpaceResponse(space);
+        if(spaceResponse == null){
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Space not correctly converted");
+        }
+        return spaceResponse;
     }
 
     @Override
@@ -53,31 +58,30 @@ public class SpaceServiceImpl implements SpaceService {
         if (space == null) {
             throw new ResourceNotFoundException("Space not found", "space", spaceId);
         }
+        User spaceOwner = space.getOwner();
+        if(!spaceOwner.getUserId().equals(((User) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUserId())){
+            throw new UnauthorizedException("You are not the owner of this space");
+        }
+
         if (editSpaceRequest.getSpaceName() != null){
             space.setSpaceName(editSpaceRequest.getSpaceName());
         }
         if (editSpaceRequest.getSpaceLocation() != null){
             space.setSpaceLocation(editSpaceRequest.getSpaceLocation());
         }
-        if (editSpaceRequest.getSpaceSize() != 0){
+        if (editSpaceRequest.getSpaceSize() > 0 ){
             space.setSpaceSize(editSpaceRequest.getSpaceSize());
         }
-        if (editSpaceRequest.getSpacePrice() != 0){
+        if (editSpaceRequest.getSpacePrice() > 0){
             space.setSpacePrice(editSpaceRequest.getSpacePrice());
         }
         if(editSpaceRequest.getSpaceDescription() != null){
             space.setSpaceDescription(editSpaceRequest.getSpaceDescription());
         }
-        if(editSpaceRequest.getOwnerId() != null){
-            Optional<User> Owner = userRepository.findByUserId(editSpaceRequest.getOwnerId());
-            if(Owner.isPresent()){
-                space.setOwner(Owner.get());
-            }
-            else{
-                throw new ResourceNotFoundException("Owner not found", "owner", editSpaceRequest.getOwnerId());
-            }
+        var result = spaceRepository.save(space);
+        if(result == null){
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Space not updated");
         }
-        spaceRepository.save(space);
         return ObjectMapper.mapSpaceToSpaceResponse(space);
     }
 
@@ -89,36 +93,19 @@ public class SpaceServiceImpl implements SpaceService {
         if (space==null) {
             throw new ResourceNotFoundException("Space not found", "space", id);
         }
-        long deleted = spaceRepository.deleteBySpaceId(id);
-        if (deleted == 0) {
+        if(!space.getOwner().getUserId().equals(((User) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUserId())){
+            throw new UnauthorizedException("You are not the owner of this space");
+        }
+        int deleted = spaceRepository.deleteBySpaceId(id);
+        if (deleted == 0 || spaceRepository.findBySpaceId(id).isPresent()) {
             throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "space not deleted");
         }
         return ObjectMapper.mapSpaceToSpaceResponse(space);
     }
     @Override
     public List<SpaceResponse> searchSpaces(SpaceFilter filter) {
-        List<Space> spaces = spaceRepository.findBySpaceSizeLessThanEqualAndSpaceSizeGreaterThanEqualAndSpacePriceLessThanEqualAndSpacePriceGreaterThanEqual(
-                filter.getSpaceSizeUpperBound(), filter.getSpaceSizeLowerBound(), filter.getSpacePriceUpperBound(), filter.getSpacePriceLowerBound()
-        );
-        if(filter.getSpaceName()!=null){
-            spaces.removeIf(space -> !space.getSpaceName().equals(filter.getSpaceName()));
-        }
-        if(filter.getSpaceLocation()!=null){
-            spaces.removeIf(space -> !space.getSpaceLocation().equals(filter.getSpaceLocation()));
-        }
-        if(filter.getSpaceType()!=null){
-            spaces.removeIf(space -> !space.getSpaceType().equals(filter.getSpaceType()));
-        }
-        if(filter.getAvailability()!=null){
-            spaces.removeIf(space -> !space.getAvailibility().equals(filter.getAvailability()));
-        }
-        if(filter.getOwnerId()!=null){
-            spaces.removeIf(space -> !space.getOwner().getUserId().equals(filter.getOwnerId()));
-        }
-        spaces.removeIf(space -> !checkSpaceAvailability(space, filter.getStartDate(), filter.getEndDate()));
-        return spaces.stream().map(ObjectMapper::mapSpaceToSpaceResponse).toList();
+        return doFilter(filter , Optional.empty()).stream().map(ObjectMapper::mapSpaceToSpaceResponse).toList();
     }
-
     private boolean checkSpaceAvailability(Space space, Date startDate, Date endDate) {
         List<Booking> bookings = bookingRepository.findBySpace_SpaceId(space.getSpaceId());
         for (Booking booking : bookings) {
@@ -149,18 +136,56 @@ public class SpaceServiceImpl implements SpaceService {
         spaceRepository.save(space);
         return ObjectMapper.mapSpaceToSpaceResponse(space);
     }
-
     @Override
-    public List<SpaceResponse> getMySpaces() {
+    public List<SpaceResponse> getMySpaces(SpaceFilter filter) {
         User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        if(user == null){
-            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "user not log in");
-        }
-        return spaceRepository.findByOwner_UserId(user.getUserId()).stream().map(ObjectMapper::mapSpaceToSpaceResponse).toList();
+        return doFilter(filter , Optional.ofNullable(user.getUserId())).stream().map(ObjectMapper::mapSpaceToSpaceResponse).toList();
     }
 
     @Override
     public List<SpaceResponse> getAllSpaces() {
         return spaceRepository.findAll().stream().map(ObjectMapper::mapSpaceToSpaceResponse).toList();
+    }
+
+    @Override
+    public SpaceBookedDates getBookedDates(String spaceId) {
+        Optional<Space> spaceOpt = spaceRepository.findBySpaceId(spaceId);
+        Space space = spaceOpt.orElse(null);
+        if (space == null) {
+            throw new ResourceNotFoundException("Space not found", "space", spaceId);
+        }
+        List<Booking> bookings = bookingRepository.findBySpace_SpaceId(spaceId);
+        SpaceBookedDates spaceBookedDates = new SpaceBookedDates();
+        spaceBookedDates.setBookedDates(ObjectMapper.mapBookingsToBookedDates(bookings));
+        return spaceBookedDates;
+    }
+
+    public Boolean checkAvailabilityForBooking(String spaceId , Date startDate , Date endDate) {
+        Optional<Space> spaceOpt = spaceRepository.findBySpaceId(spaceId);
+        Space space = spaceOpt.orElse(null);
+        if (space == null) {
+            throw new ResourceNotFoundException("Space not found", "space", spaceId);
+        }
+        return checkSpaceAvailability(space, startDate, endDate);
+    }
+    private List<Space> doFilter(SpaceFilter filter, Optional<String> userId) {
+        List<Space> spaces = spaceRepository.findSpacesByPriceRangeAndSizeLimitForOwner(filter.getSpacePriceUpperBound(), filter.getSpacePriceLowerBound() ,
+                filter.getSpacePriceUpperBound(), filter.getSpaceSizeLowerBound(), userId.orElse(null));
+        if (filter.getSpaceName() != null) {
+            spaces.removeIf(space -> !space.getSpaceName().contains(filter.getSpaceName()));
+        }
+        if(filter.getSpaceLocation()!=null){
+            spaces.removeIf(space -> !space.getSpaceLocation().contains(filter.getSpaceLocation()));
+        }
+        if(filter.getSpaceType()!=null){
+            spaces.removeIf(space -> !space.getSpaceType().equals(filter.getSpaceType()));
+        }
+        if(filter.getAvailability()!=null){
+            spaces.removeIf(space -> !space.getAvailibility().equals(filter.getAvailability()));
+        }
+        if(filter.getStartDate()!=null && filter.getEndDate()!=null){
+            spaces.removeIf(space -> !checkSpaceAvailability(space, filter.getStartDate(), filter.getEndDate()));
+        }
+        return spaces;
     }
 }
